Continue trailing stop order implementation
- Reorganize `SimulatedExchange` tests
