+++ tests/unit_tests/backtest/test_backtest_exchange_trailing_stops.py
@@ -1,0 +1,365 @@
+# -------------------------------------------------------------------------------------------------
+#  Copyright (C) 2015-2022 Nautech Systems Pty Ltd. All rights reserved.
+#  https://nautechsystems.io
+#
+#  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
+#  You may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at https://www.gnu.org/licenses/lgpl-3.0.en.html
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+# -------------------------------------------------------------------------------------------------
+
+from decimal import Decimal
+
+import pytest
+
+from nautilus_trader.backtest.data.providers import TestInstrumentProvider
+from nautilus_trader.backtest.exchange import SimulatedExchange
+from nautilus_trader.backtest.execution_client import BacktestExecClient
+from nautilus_trader.backtest.models import FillModel
+from nautilus_trader.backtest.models import LatencyModel
+from nautilus_trader.common.clock import TestClock
+from nautilus_trader.common.enums import LogLevel
+from nautilus_trader.common.logging import Logger
+from nautilus_trader.config.common import ExecEngineConfig
+from nautilus_trader.config.common import RiskEngineConfig
+from nautilus_trader.data.engine import DataEngine
+from nautilus_trader.execution.engine import ExecutionEngine
+from nautilus_trader.model.currencies import USD
+from nautilus_trader.model.data.tick import QuoteTick
+from nautilus_trader.model.data.tick import TradeTick
+from nautilus_trader.model.enums import AccountType
+from nautilus_trader.model.enums import AggressorSide
+from nautilus_trader.model.enums import OMSType
+from nautilus_trader.model.enums import OrderSide
+from nautilus_trader.model.enums import TrailingOffsetType
+from nautilus_trader.model.enums import TriggerType
+from nautilus_trader.model.identifiers import TradeId
+from nautilus_trader.model.identifiers import Venue
+from nautilus_trader.model.objects import Money
+from nautilus_trader.model.objects import Price
+from nautilus_trader.model.objects import Quantity
+from nautilus_trader.msgbus.bus import MessageBus
+from nautilus_trader.portfolio.portfolio import Portfolio
+from nautilus_trader.risk.engine import RiskEngine
+from tests.test_kit.mocks.strategies import MockStrategy
+from tests.test_kit.stubs.component import TestComponentStubs
+from tests.test_kit.stubs.data import TestDataStubs
+from tests.test_kit.stubs.identifiers import TestIdStubs
+
+
+AUDUSD_SIM = TestInstrumentProvider.default_fx_ccy("AUD/USD")
+USDJPY_SIM = TestInstrumentProvider.default_fx_ccy("USD/JPY")
+
+
+class TestSimulatedExchange:
+    def setup(self):
+        # Fixture Setup
+        self.clock = TestClock()
+        self.logger = Logger(
+            clock=self.clock,
+            level_stdout=LogLevel.DEBUG,
+        )
+
+        self.trader_id = TestIdStubs.trader_id()
+
+        self.msgbus = MessageBus(
+            trader_id=self.trader_id,
+            clock=self.clock,
+            logger=self.logger,
+        )
+
+        self.cache = TestComponentStubs.cache()
+
+        self.portfolio = Portfolio(
+            msgbus=self.msgbus,
+            cache=self.cache,
+            clock=self.clock,
+            logger=self.logger,
+        )
+
+        self.data_engine = DataEngine(
+            msgbus=self.msgbus,
+            clock=self.clock,
+            cache=self.cache,
+            logger=self.logger,
+        )
+
+        self.exec_engine = ExecutionEngine(
+            msgbus=self.msgbus,
+            cache=self.cache,
+            clock=self.clock,
+            logger=self.logger,
+            config=ExecEngineConfig(debug=True),
+        )
+
+        self.risk_engine = RiskEngine(
+            portfolio=self.portfolio,
+            msgbus=self.msgbus,
+            cache=self.cache,
+            clock=self.clock,
+            logger=self.logger,
+            config=RiskEngineConfig(debug=True),
+        )
+
+        self.exchange = SimulatedExchange(
+            venue=Venue("SIM"),
+            oms_type=OMSType.HEDGING,
+            account_type=AccountType.MARGIN,
+            base_currency=USD,
+            starting_balances=[Money(1_000_000, USD)],
+            default_leverage=Decimal(50),
+            leverages={AUDUSD_SIM.id: Decimal(10)},
+            instruments=[USDJPY_SIM],
+            modules=[],
+            fill_model=FillModel(),
+            cache=self.cache,
+            clock=self.clock,
+            logger=self.logger,
+            latency_model=LatencyModel(0),
+        )
+
+        self.exec_client = BacktestExecClient(
+            exchange=self.exchange,
+            msgbus=self.msgbus,
+            cache=self.cache,
+            clock=self.clock,
+            logger=self.logger,
+        )
+
+        # Wire up components
+        self.exec_engine.register_client(self.exec_client)
+        self.exchange.register_client(self.exec_client)
+
+        self.cache.add_instrument(USDJPY_SIM)
+
+        # Create mock strategy
+        self.strategy = MockStrategy(bar_type=TestDataStubs.bartype_usdjpy_1min_bid())
+        self.strategy.register(
+            trader_id=self.trader_id,
+            portfolio=self.portfolio,
+            msgbus=self.msgbus,
+            cache=self.cache,
+            clock=self.clock,
+            logger=self.logger,
+        )
+
+        # Start components
+        self.exchange.reset()
+        self.data_engine.start()
+        self.exec_engine.start()
+        self.strategy.start()
+
+    @pytest.mark.parametrize(
+        "order_side, trailing_offset_type, trailing_offset, expected_trigger",
+        [
+            [
+                OrderSide.BUY,
+                TrailingOffsetType.PRICE,
+                Decimal("1.0"),
+                Price.from_str("15.0"),
+            ],
+            [
+                OrderSide.SELL,
+                TrailingOffsetType.PRICE,
+                Decimal("1.0"),
+                Price.from_str("12.0"),
+            ],
+        ],
+    )
+    def test_trailing_stop_market_order_bid_ask_with_no_trigger_updates_order(
+        self,
+        order_side,
+        trailing_offset_type,
+        trailing_offset,
+        expected_trigger,
+    ):
+        # Arrange: Prepare market
+        quote = QuoteTick(
+            instrument_id=USDJPY_SIM.id,
+            bid=Price.from_str("13.0"),
+            ask=Price.from_str("14.0"),
+            bid_size=Quantity.from_int(1_000_000),
+            ask_size=Quantity.from_int(1_000_000),
+            ts_event=0,
+            ts_init=0,
+        )
+        self.exchange.process_quote_tick(quote)
+        self.data_engine.process(quote)
+        self.portfolio.update_quote_tick(quote)
+
+        trailing_stop = self.strategy.order_factory.trailing_stop_market(
+            instrument_id=USDJPY_SIM.id,
+            order_side=order_side,
+            quantity=Quantity.from_int(200000),
+            trailing_offset_type=trailing_offset_type,
+            trailing_offset=trailing_offset,
+            trigger_type=TriggerType.BID_ASK,  # <-- BID_ASK
+        )
+        self.strategy.submit_order(trailing_stop)
+        self.exchange.process(0)
+
+        # Assert
+        assert trailing_stop.trigger_price == expected_trigger
+
+    @pytest.mark.parametrize(
+        "order_side, trailing_offset_type, trailing_offset, expected_trigger",
+        [
+            [
+                OrderSide.BUY,
+                TrailingOffsetType.PRICE,
+                Decimal("1.0"),
+                Price.from_str("15.0"),
+            ],
+            [
+                OrderSide.SELL,
+                TrailingOffsetType.PRICE,
+                Decimal("1.0"),
+                Price.from_str("13.0"),
+            ],
+        ],
+    )
+    def test_trailing_stop_market_order_last_with_no_trigger_updates_order(
+        self,
+        order_side,
+        trailing_offset_type,
+        trailing_offset,
+        expected_trigger,
+    ):
+        trade = TradeTick(
+            instrument_id=USDJPY_SIM.id,
+            price=Price.from_str("14.0"),
+            size=Quantity.from_int(1),
+            aggressor_side=AggressorSide.BUY,
+            trade_id=TradeId("123456"),
+            ts_event=0,
+            ts_init=0,
+        )
+        self.exchange.process_trade_tick(trade)
+        self.data_engine.process(trade)
+
+        trailing_stop = self.strategy.order_factory.trailing_stop_market(
+            instrument_id=USDJPY_SIM.id,
+            order_side=order_side,
+            quantity=Quantity.from_int(200000),
+            trailing_offset_type=trailing_offset_type,
+            trailing_offset=trailing_offset,
+            trigger_type=TriggerType.LAST,  # <-- LAST
+        )
+        self.strategy.submit_order(trailing_stop)
+        self.exchange.process(0)
+
+        # Assert
+        assert trailing_stop.trigger_price == expected_trigger
+
+    def test_trailing_stop_market_order_buy_bid_ask_price_when_offset_activated_updates_order(self):
+        # Arrange: Prepare market
+        tick = QuoteTick(
+            instrument_id=USDJPY_SIM.id,
+            bid=Price.from_str("13.0"),
+            ask=Price.from_str("14.0"),
+            bid_size=Quantity.from_int(1_000_000),
+            ask_size=Quantity.from_int(1_000_000),
+            ts_event=0,
+            ts_init=0,
+        )
+        self.exchange.process_quote_tick(tick)
+        self.data_engine.process(tick)
+        self.portfolio.update_quote_tick(tick)
+
+        trailing_stop = self.strategy.order_factory.trailing_stop_market(
+            instrument_id=USDJPY_SIM.id,
+            order_side=OrderSide.BUY,
+            quantity=Quantity.from_int(200000),
+            trigger_price=Price.from_str("15.0"),
+            trailing_offset_type=TrailingOffsetType.PRICE,
+            trailing_offset=Decimal("1.0"),
+            trigger_type=TriggerType.BID_ASK,
+        )
+        self.strategy.submit_order(trailing_stop)
+        self.exchange.process(0)
+
+        tick = QuoteTick(
+            instrument_id=USDJPY_SIM.id,
+            bid=Price.from_str("12.0"),
+            ask=Price.from_str("13.0"),
+            bid_size=Quantity.from_int(1_000_000),
+            ask_size=Quantity.from_int(1_000_000),
+            ts_event=0,
+            ts_init=0,
+        )
+        self.exchange.process_quote_tick(tick)
+
+        # Market moves against trailing stop (should not update)
+        tick = QuoteTick(
+            instrument_id=USDJPY_SIM.id,
+            bid=Price.from_str("12.5"),
+            ask=Price.from_str("13.5"),
+            bid_size=Quantity.from_int(1_000_000),
+            ask_size=Quantity.from_int(1_000_000),
+            ts_event=0,
+            ts_init=0,
+        )
+        self.exchange.process_quote_tick(tick)
+
+        # Assert
+        assert trailing_stop.trigger_price == Price.from_str("14.0")
+
+    def test_trailing_stop_market_order_sell_bid_ask_price_when_offset_activated_updates_order(
+        self,
+    ):
+        # Arrange: Prepare market
+        tick = QuoteTick(
+            instrument_id=USDJPY_SIM.id,
+            bid=Price.from_str("13.0"),
+            ask=Price.from_str("14.0"),
+            bid_size=Quantity.from_int(1_000_000),
+            ask_size=Quantity.from_int(1_000_000),
+            ts_event=0,
+            ts_init=0,
+        )
+        self.exchange.process_quote_tick(tick)
+        self.data_engine.process(tick)
+        self.portfolio.update_quote_tick(tick)
+
+        trailing_stop = self.strategy.order_factory.trailing_stop_market(
+            instrument_id=USDJPY_SIM.id,
+            order_side=OrderSide.SELL,
+            quantity=Quantity.from_int(200000),
+            trigger_price=Price.from_str("12.0"),
+            trailing_offset_type=TrailingOffsetType.PRICE,
+            trailing_offset=Decimal("1.0"),
+            trigger_type=TriggerType.BID_ASK,
+        )
+        self.strategy.submit_order(trailing_stop)
+        self.exchange.process(0)
+
+        tick = QuoteTick(
+            instrument_id=USDJPY_SIM.id,
+            bid=Price.from_str("14.0"),
+            ask=Price.from_str("15.0"),
+            bid_size=Quantity.from_int(1_000_000),
+            ask_size=Quantity.from_int(1_000_000),
+            ts_event=0,
+            ts_init=0,
+        )
+        self.exchange.process_quote_tick(tick)
+
+        # Market moves against trailing stop (should not update)
+        tick = QuoteTick(
+            instrument_id=USDJPY_SIM.id,
+            bid=Price.from_str("13.5"),
+            ask=Price.from_str("14.5"),
+            bid_size=Quantity.from_int(1_000_000),
+            ask_size=Quantity.from_int(1_000_000),
+            ts_event=0,
+            ts_init=0,
+        )
+        self.exchange.process_quote_tick(tick)
+
+        # Assert
+        assert trailing_stop.trigger_price == Price.from_str("13.0")
